Get rid of unused perform_transactionally
The idea for this no longer made sense in a world where we use a
connection pool and not just a single connection. The actions would need
to have a way of feeding in connection params, which would require them
to know about connections in the first place, which would break the
abstraction of DbManager. It wouldn't be impossible to write, but the
result wouldn't be appropriately modular. Where multiple function calls
were needed to in a single transaction, it ended up making more sense to
write the combined functionality in at the pl/pgsql level and expose a
simplified api